Run examples as part of the test suite (#477)
* Little maintenance

* Almost working

* Update ExamplesTests.cs

* Working

* Remove

* Update tests.yaml

* Update tests.yaml

* Update install_toolchain.ps1

* Update install_toolchain.ps1

* Update ExamplesTests.cs

* E

* Update install_toolchain.ps1

* Update install_toolchain.ps1

* Fixes

* Update tests.yaml

* Update tests.yaml

* Update install_toolchain.ps1

* Update install_toolchain.ps1

+++ scripts/install_toolchain.ps1
@@ -5,6 +5,7 @@
 #>
 
 $ErrorActionPreference = "Stop"
+
 # Check if a path argument was passed in
 if ($args.Length -eq 0) {
     Write-Error "Please specify a path argument"
@@ -19,18 +20,23 @@ if (!(Test-Path $path)) {
 
 $toolchainPath = Join-Path -Path $path -ChildPath "Toolchain"
 
-# Remove previous toolchain
-Remove-Item -Path $toolchainPath -Recurse -ErrorAction SilentlyContinue
+# Remove previous toolchain, if it exists
+if (Test-Path $toolchainPath) {
+    Remove-Item -Path $toolchainPath -Recurse -ErrorAction SilentlyContinue
+}
+
+# Create the toolchain directory
+New-Item -ItemType Directory -Path $toolchainPath
 
 # Install the new toolchain in its place
-dotnet pack ../src/Draco.sln --output $toolchainPath
+dotnet pack $PSScriptRoot/../src/Draco.sln --output $toolchainPath
 
 # Install the project templates
 $templateProjectPath = Get-ChildItem -Path $toolchainPath -Filter "*Draco.ProjectTemplates.*.nupkg*" | ForEach-Object { $_.FullName }
 try { dotnet new uninstall Draco.ProjectTemplates --verbosity quiet } catch { }
 dotnet new install --force $templateProjectPath
 
-# We save the current location and go to the specified path
+# Go to the specified path
 Push-Location
 Set-Location $path
 
@@ -46,11 +52,11 @@ if (!(Test-Path $nugetConfigPath)) {
     <configuration>
       <packageSources>
         <clear />
-        <add key="draco" value=".\Toolchain" />
+        <add key="draco" value="./Toolchain" />
         <add key="nuget" value="https://api.nuget.org/v3/index.json" />
       </packageSources>
       <config>
-        <add key="globalPackagesFolder" value="Toolchain\GlobalPackages" />
+        <add key="globalPackagesFolder" value="Toolchain/GlobalPackages" />
       </config>
     </configuration>'
 
@@ -58,5 +64,4 @@ if (!(Test-Path $nugetConfigPath)) {
     Write-Host "Successfully created NuGet.config."
 }
 
-# Restore old location
 Pop-Location
